refactor: combine layouts into one to avoid duplication
