+++ src/templates/_common/_layout/head.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Title from '../../_common/components/title.jsx';
-import Favicons from '../../_common/includes/favicons.jsx';
-import AntiClickjack from '../../_common/includes/anti_clickjack.jsx';
-import GoogleOptimizer from '../../_common/includes/google_optimizer.jsx';
+import Title from '../components/title.jsx';
+import Favicons from '../includes/favicons.jsx';
+import AntiClickjack from '../includes/anti_clickjack.jsx';
+import GoogleOptimizer from '../includes/google_optimizer.jsx';
 
 const Head = () => (
     <head>
