+++ src/templates/_common/_layout/mobile_menu.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Li } from '../../_common/components/elements.jsx';
+import { Li } from '../components/elements.jsx';
 
 const MobileMenu = () => (
     <div className='gr-hide gr-show-m gr-show-t gr-show-p' id='mobile-menu-container'>
